change fortran code to take tau and T as command line inputs and add them to spectra input file; change spectra code to read them in and use them instead of previous hardcoded ones. all seems to work fine. next thing is to find a way to fix the lineshape parameters, which still need to be fiddled with in spectra code

+++ spectra/src/input.h
@@ -21,7 +21,7 @@
 typedef struct {
   unsigned int N;
   double T;
-  double tau;
+  unsigned int tau;
   char eigvecs_file[200], eigvals_file[200], mu_file[200],
   lambda_file[200], gamma_file[200], aw_file[200], fw_file[200],
   pop_file[200];
